+++ NEWS.md
@@ -6,6 +6,8 @@
 
 - Added support for the **qs2** package in `cache_exec()` and `cache_rds()` since the **qs** package will be deprecated.
 
+- Added `browser_print()` as a simple wrapper function to call a headless browser to print a web page to PDF/PNG/JPEG.
+
 # CHANGES IN xfun VERSION 0.54
 
 - `record()` tries not to records empty plots triggered as side-effects by functions like `grid::convertUnit()` (thanks, @yjunechoe, yihui/litedown#96). However, the underlying technique is not robust enough to deal with all cases. In case certain undesired plots still get recorded, you can filter them out via the new argument of `record()`, `dev.keep`, which can take indices to determine which plots to keep. For example, `dev.keep = -1` means to remove the first plot.